Added support for Forge mod commands to hook into permissions.
When a forge mod registers a command, it will automatically be registered
with Bukkit using the CommandModCustom class. This class is solely used to
register the command prefix, permission node, and allow us to check for
valid permissions through Bukkit. The permission node will use the class
name of the mod's command class that is being registered.